@@ -29,8 +29,8 @@
     "jquery": "^1.7.4",
     "lodash": "^4.17.4",
     "ohm-js": "^0.12.3",
-    "react": "^15.4.1",
-    "react-dom": "^15.4.2",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
     "string-pixel-width": "^1.0.7"
   },
   "devDependencies": {
